+++ addon/src/utils/portal.ts
@@ -2,6 +2,12 @@ import { getOwner } from '@ember/application';
 
 export function portalId(component: unknown): string {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const { APP } = (getOwner(component) as any).resolveRegistration('config:environment');
-  return APP.primitiveUI.portalElement;
+  const { APP, environment } = (getOwner(component) as any).resolveRegistration('config:environment');
+  const { portalElement } = APP.primitiveUI || {};
+
+  if (environment === 'test') {
+    return portalElement || APP.rootElement;
+  }
+
+  return portalElement || '#portal';
 }